+++ core/src/main/java/org/neo4j/graphalgo/core/neo4jview/GraphView.java
@@ -127,7 +127,7 @@ public class GraphView implements Graph {
     }
 
     @Override
-    public int nodeCount() {
+    public long nodeCount() {
         return nodeCount;
     }
